Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-ninja>=1.4.3",
    "dotenv>=0.9.9",
    "google-genai>=1.30.0",
]
